[feat] #20 Store/Category/User Mapping

+++ src/main/java/com/sparta/aibusinessproject/service/StoreService.java
@@ -1,8 +1,6 @@
 package com.sparta.aibusinessproject.service;
 
-import com.sparta.aibusinessproject.domain.Category;
-import com.sparta.aibusinessproject.domain.Store;
-import com.sparta.aibusinessproject.domain.StoreCategory;
+import com.sparta.aibusinessproject.domain.*;
 import com.sparta.aibusinessproject.domain.dto.StoreDto;
 import com.sparta.aibusinessproject.domain.request.*;
 import com.sparta.aibusinessproject.domain.response.StoreSearchListResponse;
@@ -13,6 +11,7 @@ import com.sparta.aibusinessproject.repository.CategoryRepository;
 import com.sparta.aibusinessproject.repository.StoreCategoryRepository;
 import com.sparta.aibusinessproject.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -25,6 +24,7 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class StoreService {
 
     private final CategoryService categoryService;
@@ -35,7 +35,13 @@ public class StoreService {
 
     // 가게 추가
     @Transactional
-    public String createOrder(StoreCreateRequest requestDto) {
+    public String createOrder(StoreCreateRequest requestDto, User user) {
+
+        // 유저 권한일때는 접근 권한 없음
+        if(user.getRole() == UserRoleEnum.CUSTOMER){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         StoreDto dto = StoreCreateRequest.toDto(requestDto);
 
         // 존재하는 가게이름인지 확인
@@ -45,7 +51,8 @@ public class StoreService {
         if(!store.isEmpty()) {
             throw new ApplicationException(ErrorCode.DUPLICATED_STORENAME);
         }
-        Store storeEntity = storeRepository.save(StoreDto.toEntity(dto));
+
+        Store storeEntity = storeRepository.save(StoreDto.toEntity(dto,user));
 
         return storeEntity.getName();
     }
@@ -68,37 +75,70 @@ public class StoreService {
     // 가게 세부 조회
     @Transactional
     public StoreSearchResponse getStoreById(UUID storeId) {
+
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new ApplicationException(ErrorCode.INVALID_STORE));
 
-            return  new StoreSearchResponse(store);
+        return  new StoreSearchResponse(store);
     }
 
     // 가게 수정
     @Transactional
-    public StoreDto update(UUID storeId, StoreUpdateRequest request) {
+    public StoreSearchResponse update(UUID storeId, StoreUpdateRequest request, User user) {
+        // 유저 권한일때는 접근 권한 없음
+        if(user.getRole() == UserRoleEnum.CUSTOMER){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new ApplicationException(ErrorCode.INVALID_STORE));
 
+        // 로그인한 유저와 가게 작성한 유저가 다를 경우
+        if(!store.getUser().getUserId().equals(user.getUserId())){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         store.update(request);
 
-        return  StoreDto.from(store);
+        return  new StoreSearchResponse(store);
     }
 
     // 가게 삭제
     @Transactional
-    public void delete(UUID storeId) {
+    public UUID delete(UUID storeId, User user) {
+        // 유저 권한일때는 접근 권한 없음
+        if(user.getRole() == UserRoleEnum.CUSTOMER){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new ApplicationException(ErrorCode.INVALID_STORE));
 
+        // 로그인한 유저와 가게 작성한 유저가 다를 경우
+        if(!store.getUser().getUserId().equals(user.getUserId())){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         storeRepository.delete(store);
+
+        return storeId;
     }
 
     // 가게에 대한 카테고리 생성
     @Transactional
-    public void createStoreCategory(UUID storeId, CategoryListCreateRequest request) {
+    public StoreSearchResponse createStoreCategory(UUID storeId, CategoryListCreateRequest request, User user) {
+        // 유저 권한일때는 접근 권한 없음
+        if(user.getRole() == UserRoleEnum.CUSTOMER){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new ApplicationException(ErrorCode.INVALID_STORE));
+
+        // 로그인한 유저와 가게 작성한 유저가 다를 경우
+        if(!store.getUser().getUserId().equals(user.getUserId())){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
 
         List<StoreCategory> storeCategories;
 
@@ -122,7 +162,7 @@ public class StoreService {
 
         // StoreCategory Entity 생성
 
-        // 이미 존재하는 가게에 대한 카테고리인지 확인
+        // 이미 가게에 대한 존재하는 카테고리인지 확인
         for(StoreCategory storeCategory : storeCategories){
             if(storeCategory.getCategory().getId().equals(categoryEntity.getId())){
                 throw new ApplicationException(ErrorCode.ALREADY_CATEGORY);
@@ -130,5 +170,49 @@ public class StoreService {
         }
         StoreCategory storeCategory = storeCategoryRepository.save(new StoreCategory(categoryEntity,store));
         store.categoryUpdate(storeCategory);
+
+
+        return  new StoreSearchResponse(store);
     }
+
+    // 가게에 대한 카테고리 삭제
+    @Transactional
+    public String deleteStoreCategory(UUID storeId, CategoryListCreateRequest request, User user) {
+        // 유저 권한일때는 접근 권한 없음
+        if(user.getRole() == UserRoleEnum.CUSTOMER){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
+        Store store = storeRepository.findById(storeId)
+                .orElseThrow(() -> new ApplicationException(ErrorCode.INVALID_STORE));
+
+        // 로그인한 유저와 가게 작성한 유저가 다를 경우
+        if(!store.getUser().getUserId().equals(user.getUserId())){
+            throw new ApplicationException(ErrorCode.ACCESS_DENIED);
+        }
+
+        List<StoreCategory> storeCategories;
+
+        // StoreCategory가 존재하는지 확인, 존재하지 않다면 에러 발생 / 존재한다면 기존 리스트 가져옴
+        if(store.getStoreCategories().isEmpty()){
+            throw new ApplicationException(ErrorCode.NOTFOUND_CATEGORY);
+        }
+
+
+        // 카테고리 값 존재유무 확인 및 삭제
+        Optional<Category> category = categoryRepository.findByName(request.categories());
+        Category categoryEntity;
+
+        if(category.isEmpty()) {
+            throw new ApplicationException(ErrorCode.INVALID_CATEGORY);
+        }else{
+            categoryEntity = category.get();
+            log.info(categoryEntity.getName() + "," + categoryEntity.getId());
+            // 해당 storeCategory 삭제
+            storeCategoryRepository.deleteByCategoryId(categoryEntity.getId());
+        }
+
+        return  request.categories()+"카테고리가 삭제되었습니다";
+    }
+
 }
